add error handling

+++ src/main/java/com/example/notes/services/NoteService.java
@@ -22,6 +22,9 @@ public class NoteService {
     }
 
     public NoteResponseDTO addNote(CreateNoteDTO createNoteDTO) {
+        if (Objects.isNull(createNoteDTO))
+            throw new IllegalArgumentException("Note is required");
+
         var noteEntity = modelMapper.map(createNoteDTO, NoteEntity.class);
 
         var savedNote = noteRepository.save(noteEntity);
@@ -38,11 +41,22 @@ public class NoteService {
     }
 
     public void deleteNote(Integer noteId) {
+        if (Objects.isNull(noteId))
+            throw new IllegalArgumentException("Note Id cannot be null");
+        if (!noteRepository.existsById(noteId))
+            throw new NoteNotFoundException("Note with id: " + noteId + " does not exist");
+
         noteRepository.deleteById(noteId);
     }
 
     public NoteResponseDTO updateNote(Integer noteId, CreateNoteDTO newNote) {
+        if (Objects.isNull(noteId) || Objects.isNull(newNote))
+            throw new IllegalArgumentException("Id and note to update is required");
+
         var note = getNoteById(noteId);
+
+        if (Objects.isNull(note))
+            throw new NoteNotFoundException("Note with id: " + noteId + " does not exist");
 
         if (Objects.nonNull(newNote.getTitle())) {
             note.setTitle(newNote.getTitle());
@@ -57,16 +71,33 @@ public class NoteService {
     }
 
     public NoteResponseDTO getNoteById(Integer noteId) {
+        if (Objects.isNull(noteId) || noteId.equals(0))
+            throw new IllegalArgumentException("Id cannot be null or zero");
+
         var noteEntity = noteRepository.findById(noteId);
+
+        if (noteEntity.isEmpty()) throw new NoteNotFoundException("Note with id: " + noteId
+                + " does not exist");
 
         return modelMapper.map(noteEntity, NoteResponseDTO.class);
     }
 
     public List<NoteResponseDTO> getAllNotesByTitlePrefix(String noteTitlePrefix) {
+        if (noteTitlePrefix.isEmpty())
+            throw new IllegalArgumentException("Prefix cannot be empty or null");
+
         var notes = noteRepository.findByTitleStartsWithIgnoreCase(noteTitlePrefix);
+
+        if (notes.isEmpty()) throw new NoteNotFoundException("Note does not exist");
 
         return notes.stream()
                 .map(note -> modelMapper.map(note, NoteResponseDTO.class))
                 .toList();
+    }
+
+    public static class NoteNotFoundException extends RuntimeException {
+        public NoteNotFoundException(String message) {
+            super(message);
+        }
     }
 }
